доработки по переезду с репо на стор

+++ migrations/sql/V1__base_tables.sql
@@ -103,7 +103,7 @@ where false;
 
 create table if not exists auth.auth_history
 (
-    id           int primary key,
+    id           bigserial primary key,
     user_        int,
     attempt_type text,
     ip           text,
